Extend the model to support validity periods for budgets.
Budgets are only valid for a certain configurable period. This allows to e.g. specify a budget 'Food' with some value for the year 2018, and increase the budget for 2019 due to higher income. This change only extends the backend's model code with this possibility. It is not yet possible to query for the valid budgets in some period. The frontend also does not yet support this change.