Add a REST handler for templates.
As the REST handler for templates is essentially the same as for expenses, we had to change the code a bit.

Handlers can now implement the 'RestHandler' interface to specify the REST methods that they can handle.

We split up the additional 'utility' methods of the Expense handler (tags, payment methods, authors) and moved it to a new handler (as these will stay for templates).

Then we adapted the ExpenseRestHandler to handle both expenses and templates.