Introduce custom budget views.
We allow to define custom views based on all or a list of budgets, either monthly or yearly. This way, users can get an overview of their budget in the way they want to. Adding the views via REST API is not yet possible, however they can e.g. be added to the file manually.